feat: expanding spec validation to openapi 3.x
