+++ .eslintrc.yml
@@ -4,6 +4,8 @@ env:
   browser: true
   node: true
   mocha: true
+parserOptions:
+  ecmaVersion: 2020
 rules:
   no-plusplus: off
   no-restricted-syntax: off
