+++ test/specs/validate-spec/invalid/2.0/array-body-no-items.yaml
