+++ test/specs/validate-spec/invalid/2.0/array-no-items.yaml
