+++ test/specs/validate-spec/invalid/2.0/array-response-body-no-items.yaml
