+++ test/specs/validate-spec/invalid/2.0/array-response-header-no-items.yaml
