+++ test/specs/validate-spec/invalid/2.0/body-and-form-params.yaml
