+++ test/specs/validate-spec/invalid/2.0/duplicate-operation-params.yaml
