+++ test/specs/validate-spec/invalid/2.0/duplicate-path-placeholders.yaml
