+++ test/specs/validate-spec/invalid/2.0/file-invalid-consumes.yaml
