+++ test/specs/validate-spec/invalid/2.0/invalid-response-code.yaml
