+++ test/specs/validate-spec/invalid/2.0/multiple-body-params.yaml
