+++ test/specs/validate-spec/invalid/2.0/multiple-operation-body-params.yaml
