+++ test/specs/validate-spec/invalid/2.0/multiple-path-body-params.yaml
