+++ test/specs/validate-spec/invalid/2.0/no-path-params.yaml
