+++ test/specs/validate-spec/invalid/2.0/path-param-no-placeholder.yaml
