+++ test/specs/validate-spec/invalid/2.0/path-placeholder-no-param.yaml
