+++ test/specs/validate-spec/invalid/2.0/required-property-not-defined-definitions.yaml
