+++ test/specs/validate-spec/invalid/3.x/array-no-items.yaml
@@ -1,0 +1,16 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users:
+    parameters:
+      - name: tags
+        in: query
+        schema:
+          type: array
+    get:
+      responses:
+        default:
+          description: hello world
