+++ test/specs/validate-spec/invalid/3.x/array-response-header-no-items.yaml
@@ -1,0 +1,18 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users:
+    get:
+      responses:
+        "default":
+          description: hello world
+          headers:
+            Content-Type:
+              schema:
+                type: string
+            Last-Modified:
+              schema:
+                type: array
