+++ test/specs/validate-spec/invalid/3.x/duplicate-operation-ids.yaml
@@ -1,0 +1,17 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users:
+    get:
+      operationId: users
+      responses:
+        default:
+          description: hello world
+    post:
+      operationId: users          # <---- duplicate
+      responses:
+        default:
+          description: hello world
