+++ test/specs/validate-spec/invalid/3.x/duplicate-operation-params.yaml
@@ -1,0 +1,31 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users/{username}:
+    get:
+      parameters:
+        - name: username           # <---- Duplicate param
+          in: path
+          required: true
+          schema:
+            type: string
+        - name: bar
+          in: header
+          required: false
+          schema:
+            type: string
+        - name: username
+          in: header
+          schema:
+            type: string
+        - name: username           # <---- Duplicate param
+          in: path
+          required: true
+          schema:
+            type: number
+      responses:
+        default:
+          description: hello world
