+++ test/specs/validate-spec/invalid/3.x/no-path-params.yaml
@@ -1,0 +1,33 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users/{username}/{foo}:        # <---- {username} and {foo} placeholders
+    parameters:                   # <---- no path params
+      - name: username
+        in: header
+        required: true
+        schema:
+          type: string
+    get:
+      parameters:                 # <---- no path params
+        - name: username
+          in: header
+          required: true
+          schema:
+            type: number
+      responses:
+        default:
+          description: hello world
+    post:
+      parameters:                 # <---- no path params
+        - name: username
+          in: header
+          required: true
+          schema:
+            type: string
+      responses:
+        default:
+          description: hello world
