+++ test/specs/validate-spec/invalid/3.x/path-placeholder-no-param.yaml
@@ -1,0 +1,33 @@
+openapi: 3.0.0
+info:
+  version: "1.0.0"
+  title: Invalid API
+
+paths:
+  /users/{username}/{foo}:        # <---- {username} and {foo} placeholders
+    parameters:
+      - name: username            # <---- "username" path param
+        in: path
+        required: true
+        schema:
+          type: string
+    get:
+      parameters:                   # <---- there's no "foo" path param
+        - name: username
+          in: path
+          required: true
+          schema:
+            type: string
+      responses:
+        default:
+          description: hello world
+    post:
+      parameters:                   # <---- there's no "foo" path param
+        - name: username
+          in: path
+          required: true
+          schema:
+            type: string
+      responses:
+        default:
+          description: hello world
