+++ test/specs/validate-spec/invalid/3.x/required-property-not-defined-components.yaml
@@ -1,0 +1,36 @@
+openapi: 3.0.0
+info:
+  version: 1.0.0
+  title: Swagger Petstore
+paths:
+  '/pet/{petId}':
+      get:
+        summary: Find pet by ID
+        parameters:
+          - name: petId
+            in: path
+            description: ID of pet to return
+            required: true
+            schema:
+              type: integer
+              format: int64
+        responses:
+          '200':
+            description: successful operation
+            content:
+              application/json:
+                schema:
+                  $ref: '#/components/schemas/Pet'
+components:
+  schemas:
+    Pet:
+      type: object
+      required:
+        - name
+        - photoUrls                        # <--- does not exist
+      properties:
+        name:
+          type: string
+          example: doggie
+        color:
+          type: string
