+++ test/specs/validate-spec/invalid/3.x/required-property-not-defined-input.yaml
@@ -1,0 +1,35 @@
+openapi: 3.0.0
+info:
+  version: 1.0.0
+  title: Swagger Petstore
+paths:
+  /pets:
+    post:
+      description: Creates a new pet in the store
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              type: object
+              properties:
+                pet:
+                  description: Pet to add to the store
+                  type: object
+                  required:
+                    - notExists                # <--- does not exist
+                  properties:
+                    name:
+                      type: string
+                    color:
+                      type: string
+      responses:
+        '200':
+          description: pet response
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  name:
+                    type: string
