+++ test/specs/validate-spec/valid/2.0/file-vendor-specific-consumes-urlencoded.yaml
