+++ test/specs/validate-spec/valid/2.0/inherited-required-properties.yaml
